requirements.txt
TA-lib
numpy
websocket-client
python-socketio
requests